Move files to more logical structure
