+++ source/node/Descriptor.js
