+++ source/node/commands/BaseCommand.js
