+++ source/node/commands/CommentCommand.js
