+++ source/node/commands/CreateEntryCommand.js
