+++ source/node/commands/DeleteEntryCommand.js
