+++ source/node/commands/DeleteEntryMetaCommand.js
