+++ source/node/commands/DeleteGroupAttributeCommand.js
