+++ source/node/commands/DeleteGroupCommand.js
