+++ source/node/commands/FormatCommand.js
