+++ source/node/commands/MoveEntryCommand.js
