+++ source/node/commands/MoveGroupCommand.js
