+++ source/node/commands/SetEntryAttributeCommand.js
