+++ source/node/commands/SetEntryMetaCommand.js
