+++ source/node/commands/SetEntryPropertyCommand.js
