+++ source/node/commands/SetGroupAttributeCommand.js
