+++ source/node/storage/MemoryStorageInterface.js
