+++ source/node/storage/StorageInterface.js
